+++ packages/yourloops/lib/auth/models/change-user-role-to-hcp-payload.model.ts
@@ -25,24 +25,12 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import * as auth0Mock from '@auth0/auth0-react'
-import { UserRoles } from '../../../lib/auth/models/enums/user-roles.enum'
-import { AuthenticatedUserMetadata } from '../../../lib/auth/models/enums/authenticated-user-metadata.enum'
+import { Consent } from './consent.model'
+import { HcpProfession } from './enums/hcp-profession.enum'
 
-export const loggedInUserId = '919b1575bad58'
-export const loggedInUserEmail = 'john.doe@example.com'
-export const getAccessTokenSilentlyMock = jest.fn()
-
-export const mockAuth0Hook = (role: UserRoles = UserRoles.hcp, userId = loggedInUserId) => {
-  (auth0Mock.useAuth0 as jest.Mock).mockReturnValue({
-    isAuthenticated: true,
-    isLoading: false,
-    user: {
-      email: loggedInUserEmail,
-      email_verified: true,
-      sub: 'auth0|' + userId,
-      [AuthenticatedUserMetadata.Roles]: [role]
-    },
-    getAccessTokenSilently: getAccessTokenSilentlyMock
-  })
+export interface ChangeUserRoleToHcpPayload {
+  termsOfUse: Consent
+  privacyPolicy: Consent
+  contactConsent: Consent
+  hcpProfession: HcpProfession
 }