+++ packages/yourloops/test/integration/pages/signup-stepper/caregiver-signup.spec.tsx
@@ -26,7 +26,7 @@
  */
 
 import { act, screen } from '@testing-library/react'
-import { loggedInUserEmail, loggedInUserId, mockAuth0Hook } from '../../mock/mockAuth0Hook'
+import { getAccessTokenSilentlyMock, loggedInUserEmail, loggedInUserId, mockAuth0Hook } from '../../mock/mockAuth0Hook'
 import { checkAccountSelectorStep, checkConsentStep, checkProfileStep, checkStepper } from '../../assert/signup-stepper'
 import { mockUserApi } from '../../mock/mockUserApi'
 import userEvent from '@testing-library/user-event'
@@ -84,5 +84,6 @@ describe('Signup stepper as caregiver', () => {
         settings: { country: CountryCodes.France }
       })
     )
+    expect(getAccessTokenSilentlyMock).toHaveBeenCalledWith({ ignoreCache: true })
   })
 })